+++ libs/wingc/src/dtsify/mod.rs
@@ -1,0 +1,466 @@
+use std::cell::RefCell;
+
+use camino::{Utf8Path, Utf8PathBuf};
+use indexmap::IndexMap;
+use itertools::Itertools;
+
+use crate::{
+	ast::*,
+	diagnostic::report_diagnostic,
+	file_graph::FileGraph,
+	files::Files,
+	jsify::codemaker::CodeMaker,
+	type_check::Types,
+	visit_context::{VisitContext, VisitorWithContext},
+	WINGSDK_ASSEMBLY_NAME,
+};
+
+const TYPE_INFLIGHT_POSTFIX: &str = "$Inflight";
+const TYPE_INTERNAL_NAMESPACE: &str = "$internal";
+const TYPE_STD: &str = "std";
+
+pub struct DTSifyContext<'a> {
+	pub visit_ctx: &'a mut VisitContext,
+	pub source_path: Option<&'a Utf8Path>,
+}
+
+pub struct DTSifier<'a> {
+	preflight_file_map: &'a IndexMap<Utf8PathBuf, String>,
+	source_file_graph: &'a FileGraph,
+	pub types: &'a mut Types,
+	pub output_files: RefCell<Files>,
+}
+
+impl VisitorWithContext for DTSifyContext<'_> {
+	fn ctx(&mut self) -> &mut VisitContext {
+		&mut self.visit_ctx
+	}
+}
+
+impl<'a> DTSifier<'a> {
+	pub fn new(
+		types: &'a mut Types,
+		preflight_file_map: &'a IndexMap<Utf8PathBuf, String>,
+		source_file_graph: &'a FileGraph,
+	) -> Self {
+		Self {
+			preflight_file_map,
+			source_file_graph,
+			output_files: RefCell::new(Files::new()),
+			types,
+		}
+	}
+
+	pub fn dtsify(&self, source_path: &Utf8Path, scope: &Scope) {
+		let visit_ctx = &mut VisitContext::new();
+		let mut dtsify_ctx = DTSifyContext {
+			visit_ctx,
+			source_path: Some(source_path),
+		};
+		dtsify_ctx.visit_ctx.push_env(self.types.get_scope_env(&scope));
+
+		let mut dts = CodeMaker::default();
+
+		if source_path.is_dir() {
+			let directory_children = self.source_file_graph.dependencies_of(source_path);
+
+			for file in directory_children {
+				let preflight_file_name = self.preflight_file_map.get(file).expect("no emitted JS file found");
+				if file.is_dir() {
+					let directory_name = file.file_stem().unwrap();
+					dts.line(format!("export * as {directory_name} from \"./{preflight_file_name}\""));
+				} else {
+					dts.line(format!("export * from \"./{preflight_file_name}\""));
+				}
+			}
+		} else {
+			dts.line(format!(
+				"import * as {TYPE_INTERNAL_NAMESPACE} from \"{WINGSDK_ASSEMBLY_NAME}/lib/core/types\""
+			));
+			dts.line(format!("import {{ {TYPE_STD} }} from \"{WINGSDK_ASSEMBLY_NAME}\""));
+		}
+
+		for statement in &scope.statements {
+			dts.add_code(self.dtsify_statement(&mut dtsify_ctx, statement));
+		}
+
+		let mut dts_file_name = self.preflight_file_map.get(source_path).unwrap().clone();
+		assert!(dts_file_name.ends_with(".js"));
+
+		dts_file_name.replace_range((dts_file_name.len() - 3).., ".d.ts");
+
+		match self.output_files.borrow_mut().add_file(dts_file_name, dts.to_string()) {
+			Ok(()) => {}
+			Err(err) => report_diagnostic(err.into()),
+		}
+	}
+
+	fn dtsify_function_signature(&self, ctx: &mut DTSifyContext, f: &FunctionSignature, as_inflight: bool) -> String {
+		let mut args = vec![];
+
+		let is_inflight = matches!(f.phase, Phase::Inflight);
+
+		for (i, arg) in f.parameters.iter().enumerate() {
+			let arg_name = if arg.name.name.is_empty() {
+				// function type annotations don't always have names
+				format!("arg{}", i)
+			} else {
+				arg.name.name.clone()
+			};
+			args.push(format!(
+				"{arg_name}: {}",
+				self.dtsify_type_annotation(ctx, &arg.type_annotation, as_inflight)
+			));
+		}
+
+		let return_type = self.dtsify_type_annotation(ctx, &f.return_type, as_inflight);
+		let return_type = if is_inflight {
+			format!("Promise<{return_type}>")
+		} else {
+			return_type
+		};
+
+		let func = format!("({}) => {return_type}", args.join(", "),);
+
+		if is_inflight && !as_inflight {
+			// this is the preflight side of an inflight function
+			format!("{TYPE_INTERNAL_NAMESPACE}.Inflight<{func}>")
+		} else {
+			func
+		}
+	}
+
+	fn dtsify_interface(&self, ctx: &mut DTSifyContext, interface: &Interface, as_inflight: bool) -> CodeMaker {
+		let mut code = CodeMaker::default();
+		let interface_name = if as_inflight {
+			format!("{}{TYPE_INFLIGHT_POSTFIX}", &interface.name.name)
+		} else {
+			interface.name.name.to_string()
+		};
+
+		code.line(format!("export interface {interface_name}"));
+
+		if !interface.extends.is_empty() {
+			code.append(" extends ");
+			code.append(
+				&interface
+					.extends
+					.iter()
+					.map(|udt| {
+						if as_inflight {
+							format!("{udt}{TYPE_INFLIGHT_POSTFIX}")
+						} else {
+							udt.to_string()
+						}
+					})
+					.join(", "),
+			);
+		}
+
+		code.open("{");
+
+		for method in &interface.methods {
+			if (as_inflight && matches!(method.1.phase, Phase::Preflight))
+				|| (!as_inflight && matches!(method.1.phase, Phase::Inflight))
+			{
+				continue;
+			}
+			code.line(format!(
+				"readonly {}: {};",
+				method.0.name,
+				self.dtsify_function_signature(ctx, &method.1, as_inflight)
+			));
+		}
+
+		code.close("}");
+
+		code
+	}
+
+	fn dtsify_class(&self, ctx: &mut DTSifyContext, class: &Class, as_inflight: bool) -> CodeMaker {
+		let mut code = CodeMaker::default();
+		let class_name = if as_inflight {
+			format!("{}{TYPE_INFLIGHT_POSTFIX}", class.name)
+		} else {
+			class.name.name.to_string()
+		};
+
+		code.line("export class ");
+
+		code.append(&class_name);
+		if let Some(parent) = &class.parent {
+			code.append(" extends ");
+			if as_inflight {
+				code.append(format!("{parent}{TYPE_INFLIGHT_POSTFIX}"));
+			} else {
+				code.append(parent.to_string());
+			}
+		} else if !as_inflight && matches!(class.phase, Phase::Preflight) {
+			code.append(format!(" extends {TYPE_STD}.Resource"));
+		}
+
+		if !class.implements.is_empty() {
+			code.append(" implements ");
+			if as_inflight {
+				code.append(
+					&class
+						.implements
+						.iter()
+						.map(|udt| format!("{udt}{TYPE_INFLIGHT_POSTFIX}"))
+						.join(", "),
+				);
+			} else {
+				code.append(&class.implements.iter().map(|udt| udt.to_string()).join(", "));
+			}
+		}
+
+		code.open("{");
+
+		let constructor_params = self.dtsify_parameters(ctx, &class.initializer.signature.parameters);
+		if matches!(class.phase, Phase::Preflight) {
+			if as_inflight {
+				code.line(format!("constructor({constructor_params});"));
+			} else {
+				// add implicit scope and id
+				code.line(format!(
+					"constructor(scope: {TYPE_INTERNAL_NAMESPACE}.Construct, id: string"
+				));
+
+				if constructor_params.is_empty() {
+					code.append(");");
+				} else {
+					code.append(format!(", {constructor_params});"));
+				}
+			}
+		} else {
+			code.line(format!("constructor({constructor_params});"));
+		}
+
+		if !as_inflight && matches!(class.phase, Phase::Preflight) {
+			// Emit special preflight marks to tie the inflight side together
+			let inflight_class_name = format!("{class_name}{TYPE_INFLIGHT_POSTFIX}");
+			code.line(format!(
+				"[{TYPE_INTERNAL_NAMESPACE}.INFLIGHT_SYMBOL]?: {inflight_class_name};"
+			));
+			code.line(format!(
+				"_supportedOps(): {TYPE_INTERNAL_NAMESPACE}.OperationsOf<{inflight_class_name}>;"
+			));
+		}
+
+		for field in class
+			.fields
+			.iter()
+			.filter(|f| matches!(f.access, AccessModifier::Public))
+		{
+			if (as_inflight && matches!(field.phase, Phase::Preflight))
+				|| (!as_inflight && matches!(field.phase, Phase::Inflight))
+			{
+				continue;
+			}
+
+			code.line(format!(
+				"{}: {};",
+				field.name,
+				self.dtsify_type_annotation(ctx, &field.member_type, true)
+			));
+		}
+		for method in class
+			.methods
+			.iter()
+			.filter(|f| matches!(f.1.access, AccessModifier::Public))
+		{
+			if (as_inflight && matches!(method.1.signature.phase, Phase::Preflight))
+				|| (!as_inflight && matches!(method.1.signature.phase, Phase::Inflight))
+			{
+				continue;
+			}
+
+			code.line(format!(
+				"{}{}: {};",
+				if method.1.is_static { "static " } else { "" },
+				method.0.name,
+				self.dtsify_function_signature(ctx, &method.1.signature, as_inflight)
+			));
+		}
+
+		code.close("}");
+		code
+	}
+
+	fn dtsify_statement(&self, ctx: &mut DTSifyContext, stmt: &Stmt) -> CodeMaker {
+		let mut code = CodeMaker::default();
+		match &stmt.kind {
+			StmtKind::Interface(interface) => {
+				code.line(self.dtsify_interface(ctx, interface, false));
+				code.line(self.dtsify_interface(ctx, interface, true));
+			}
+			StmtKind::Struct {
+				name,
+				extends,
+				fields,
+				access: _,
+			} => {
+				if !extends.is_empty() {
+					code.open(format!(
+						"export interface {} extends {} {{",
+						name.name,
+						extends.iter().map(|udt| udt.to_string()).join(", ")
+					));
+				} else {
+					code.open(format!("export interface {} {{", name.name));
+				}
+
+				for field in fields {
+					let is_optional = matches!(field.member_type.kind, TypeAnnotationKind::Optional(_));
+
+					code.line(format!(
+						"readonly {}{}: {};",
+						field.name,
+						if is_optional { "?" } else { "" },
+						self.dtsify_type_annotation(ctx, &field.member_type, false)
+					));
+				}
+				code.close("}");
+			}
+			StmtKind::Enum {
+				name,
+				values,
+				access: _,
+			} => {
+				code.open(format!("export enum {} {{", name.name));
+				for value in values {
+					code.line(format!("{},", value.name));
+				}
+				code.close("}");
+			}
+			StmtKind::Bring { source, identifier } => {
+				let identifier = identifier.as_ref().map(|i| i.name.clone()).unwrap_or("".to_string());
+
+				match source {
+					BringSource::BuiltinModule(sym) => code.line(format!("import {{ {sym} }} from \"{WINGSDK_ASSEMBLY_NAME}\"")),
+					BringSource::TrustedModule(sym, path) => code.line(format!("import * as {sym} from \"{path}\"")),
+					BringSource::WingLibrary(sym, path) => code.line(format!("import * as {sym} from \"{path}\"")),
+					BringSource::JsiiModule(sym) => code.line(format!("import * as {identifier} from \"{sym}\"")),
+					BringSource::WingFile(sym) => {
+						let preflight_file_name = self
+							.preflight_file_map
+							.get(Utf8Path::new(&sym.name))
+							.unwrap()
+							.replace(".js", "");
+						code.line(format!("import * as {identifier} from \"./{preflight_file_name}\";"))
+					}
+					BringSource::Directory(sym) => {
+						let preflight_file_name = self
+							.preflight_file_map
+							.get(Utf8Path::new(&sym.name))
+							.unwrap()
+							.replace(".js", "");
+						code.line(format!("import * as {identifier} from \"./{preflight_file_name}\";"))
+					}
+				}
+			}
+			StmtKind::Class(class) => {
+				code.line(self.dtsify_class(ctx, class, false));
+				if class.phase == Phase::Preflight {
+					code.line(self.dtsify_class(ctx, class, true));
+				}
+			}
+
+			// No need to emit anything for these
+			StmtKind::SuperConstructor { .. }
+			| StmtKind::Let { .. }
+			| StmtKind::ForLoop { .. }
+			| StmtKind::While { .. }
+			| StmtKind::IfLet(_)
+			| StmtKind::If { .. }
+			| StmtKind::Break
+			| StmtKind::Continue
+			| StmtKind::Return(_)
+			| StmtKind::Throw(_)
+			| StmtKind::Expression(_)
+			| StmtKind::Assignment { .. }
+			| StmtKind::Scope(_)
+			| StmtKind::TryCatch { .. }
+			| StmtKind::CompilerDebugEnv => {}
+		}
+
+		code
+	}
+
+	fn dtsify_parameters(&self, ctx: &mut DTSifyContext, arg_list: &Vec<FunctionParameter>) -> String {
+		let mut args = vec![];
+		for arg in arg_list {
+			args.push(format!(
+				"{}: {}",
+				arg.name,
+				self.dtsify_type_annotation(ctx, &arg.type_annotation, false)
+			));
+		}
+		args.join(", ")
+	}
+
+	fn dtsify_type_annotation(&self, ctx: &mut DTSifyContext, typ: &TypeAnnotation, ignore_phase: bool) -> String {
+		match &typ.kind {
+			TypeAnnotationKind::Inferred => panic!("Should not have any inferred types"),
+			TypeAnnotationKind::Number => "number".to_string(),
+			TypeAnnotationKind::String => "string".to_string(),
+			TypeAnnotationKind::Bool => "boolean".to_string(),
+			TypeAnnotationKind::Void => "void".to_string(),
+			TypeAnnotationKind::Json => format!("Readonly<{TYPE_INTERNAL_NAMESPACE}.Json>"),
+			TypeAnnotationKind::MutJson => format!("{TYPE_INTERNAL_NAMESPACE}.Json"),
+			TypeAnnotationKind::Duration => format!("{TYPE_STD}.Duration"),
+			TypeAnnotationKind::Optional(t) => {
+				format!("({}) | undefined", self.dtsify_type_annotation(ctx, &t, ignore_phase))
+			}
+			TypeAnnotationKind::Array(t) => format!("(readonly ({})[])", self.dtsify_type_annotation(ctx, &t, ignore_phase)),
+			TypeAnnotationKind::MutArray(t) => format!("({})[]", self.dtsify_type_annotation(ctx, &t, ignore_phase)),
+			TypeAnnotationKind::Map(t) => format!(
+				"Readonly<Record<string, {}>>",
+				self.dtsify_type_annotation(ctx, &t, ignore_phase)
+			),
+			TypeAnnotationKind::MutMap(t) => {
+				format!("Record<string, {}>", self.dtsify_type_annotation(ctx, &t, ignore_phase))
+			}
+			TypeAnnotationKind::Set(t) => format!("Readonly<Set<{}>>", self.dtsify_type_annotation(ctx, &t, ignore_phase)),
+			TypeAnnotationKind::MutSet(t) => format!("Set<{}>", self.dtsify_type_annotation(ctx, &t, ignore_phase)),
+			TypeAnnotationKind::Function(f) => self.dtsify_function_signature(ctx, f, ignore_phase),
+			TypeAnnotationKind::UserDefined(udt) => udt.to_string(),
+		}
+	}
+}
+
+#[test]
+fn types_declarations() {
+	assert_compile_dir!(
+		r#"
+    pub struct Struct {
+      n: num;
+      d: Array<duration>;
+      j: Json;
+    }
+    
+    pub interface Interface {
+      method(s: Struct): str;
+      inflight inflightMethod(): str;
+    }
+    
+    pub interface ClassInterface {
+      addHandler(handler: inflight (str): str);
+      inflight bar();
+    }
+    
+    pub class ParentClass impl ClassInterface {
+      pub static inflight static_method() {}
+    
+      inflight foo() {}
+      pub inflight bar() {}
+    
+      pub addHandler(handler: inflight (str): str) {}
+    }
+    
+    pub class Child extends ParentClass impl ClassInterface {
+      
+    }
+    "#
+	);
+}
